Eliminate ambiguous Checkstyle checks
Remove checks that differ primarily in their scope. Since these checks
include not only elements of the given visibility but also elements that
are more visible, the Checkstyle report would sometimes contain the
result of one check, sometimes that of the other check, which leads to
different issue fingerprints and thus makes it impossible for e.g.
GitLab to identify which issues are actually the same.

Approved-by: Leonard Schüngel <leonard.schuengel@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>