Use node's allowed deviation as a minimum
When extending the allowed deviation radius for the first node on a
route to include the vehicle position, ensure that it is really only
extended but never reduced. I.e. always use the deviation of the node as
the minimum allowed deviation, even if the vehicle is closer than that
to the node's coordinates.

Approved-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>