Allow "null" values for optional JSON properties
The VDA5050 specification does not describe how to handle "null" values
for optional properties in messages. Generally, depending on the
implementation, a value of "null" for an optional property could
indicate that the property has been omitted / explicitly not set.
Allowing "null" values for optional properties makes the vehicle drivers
more robust and work with vehicles that that use "null" values instead
of omitting the property entirely from JSON messages.

While here, fix the follwoing:

* Fix some of the fields that are optional according to the VDA5050
  specification and correctly treat them as such in the driver
  implementations.
* Remove some fields (from messages) that have been removed with
  VDA5050 2.0.

Co-authored-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>
Co-authored-by: Sebastian Bonna <sebastian.bonna@iml.fraunhofer.de>
Merged-by: Martin Grzenia <martin.grzenia@iml.fraunhofer.de>