Adjust example configuration
The example configuration was updated at the configuration entry
'commadapter.vehicle.vda5050.onOpModeChangeDoWithdrawOrder', to ensure
a transport order, currently processed, gets withdrawn when switching
to the manual mode as the VDA5050 specification states that the vehicle
does not receive drive orders/actions in the manual mode anyways.
Therefore it is recommended to not change this configuration entry.

Co-authored-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>
Merged-by: Stefan Walter <stefan.walter@iml.fraunhofer.de>